+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [x] Fix DM message sizing and layout
 - [x] Fix bug with relay limit mutable state
 - [x] Open relay detail page for relay links in content
+- [x] Include mutes in WoT score calculation
 
 # 0.3.11